refactor/organizing a bit
