+++ src/components/YuzuHeader/YuzuHeader.tsx
@@ -1,0 +1,14 @@
+const YuzuHeader = () => {
+  return (
+    <div className="navigation flex-shrink fixed w-full">
+      <h2>
+        🍋 <span className="">Yuzu</span>{" "}
+        <span className="text-amber-600 font-sans font-normal text-xl">
+          Finance
+        </span>
+      </h2>
+    </div>
+  );
+};
+
+export default YuzuHeader;
