+++ src/queries/index.ts
@@ -1,0 +1,2 @@
+export { WatchListQuery } from "./watchListQuery";
+export { usEquities, crypto, forex } from "./topSymbols";
